split event_dispatcher
